:construction: LegacyEmrCluster and EmrCluster
